신종하 : 플레이어 이동방향 설정 구문 완료

+++ GameEngineBase/GameEngineTime.cpp
@@ -58,7 +58,7 @@ void GameEngineTime::TimeCheck()
 
 	for (auto& Event : AllEvent_)
 	{
-		Event->Time_ -= deltaTime_;
+		Event->Time_ -= static_cast<float>(deltaTime_);
 		if (0 >= Event->Time_)
 		{
 			Event->Event_();
